chimera: Optimize inode creation with stored procedure
Motivation:

Creating name space entries is a time critical operation in Chimera
and an operation that touches several tables. Currently several
round trips to the database are required to create an inode, create
an entry in a directory and update the inode of the parent directory.

Modification:

Create a 'createInode' stored procedure that create an entry in
t_inodes and in t_dirs, and updates the parent inode in t_inodes.

Result:

Rather than three round trips there is only one. This compensates
for the extra round-trip introduced in patch 8463 as well as 8462.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8494/