chimera: Close datasource on shutdown
Motivation:

A recent refactoring caused the close of the data source to be lost. This in
turn causes problems when using HSQLDB as the last changes are not necessarily
persistent at the time of shutdown.

Modification:

Close the data source when the file system is closed.

Result:

system-test is able to persist changes to disk.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8503/