chimera: Use auto-increment field as itagid
Motivation:

Tags use a 36 character UUID as a parimary key. A text field is a relatively
expensive primary key, both in terms of CPU consumption when comparing values
and in terms of storage in the table and the index.

The tag id is a purely internal concept not exposed in the Chimera API.

Motivation:

Changes the itagid type to an auto-increment bigint value. Upon upgrade, the
existing entries are converted. The Chimera code is updated accordingly.

Downgrade code is available, although the original key values cannot be
restored.

Result:

At NDGF, the physical storage required for t_tags shrinks from 19 GB to 15 GB
using this patch. The disk space is not the siginificant aspect though - rather
the reduced size frees up memory to cache more entries.

The original use case of being able to merge two Chimera instances can still be
supported with a little extra work: When merging two databases, A and B, one
first determines the maximum itagid in A and then increments all itagids in B
with that value. One can then merge B into A and update the sequence object to
the new maximum itagid. To facilitate this use case, the foreign key constraint
between t_tags and t_tags_inodes has been marked as cascade on update.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8509/