chimera: Resolve performance regression in directory deletion
Motivation:

Upon deleting directories we remove the tags associated with the directory.  To
do this we have to check whether other directories refer to the same tags.
PostgreSQL correctly translates the SQL expression we use to this to an
anti-join, but the anti-join is strangely expensive when the right side of the
join has many entries.

Modification:

Alter the expression such that the sub-query is a constant. This is equivalent
for our use-case and is enough to make PostgreSQL use a faster query plan.

Result:

Resolves a performance regression in directory removal.

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: https://rb.dcache.org/r/8506/