pnfsmanager: Fix regression in cancelling uploads
Motivation:

The recent refactoring of Chimera transaction handling broke the cancel-upload
operation of PnfsManager: The problem is that the name space provider relies on
an exception as part of its expected work-flow and Chimera marks the current
transaction as rollback online, causing the entire operation to fail instead.

An easy solution to support such work-flows would be to mark each Chimera call
as a nested transaction, however with the PostgreSQL JDBC driver this appears
to add another round-trip to the server and is thus undesirable.

Modification:

Instead this patch reworks the recursively directory deletion code to avoid
the exception in non-failure flows. The new code also avoid some redundant
inode lookups.

Result:

The regression is resolved and the new code has slightly better performance
on both commit and cancellation. Fixes #1826.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8518/