spacemanager: Fix race condition leading to leaked reservation entries
Motivation:

Space manager suffers from a race causing it to leak reservation entries
upon upload failures. When it occurs, it leaves behind a reservation entry
even though the file has been deleted.

The race can be triggered in several ways. The likely scenario reported by a
site is when an SRM upload expires and the SRM deletes the name space entry and
upload directory for the file in combination with the client itself aborting
the FTP upload. In this case the upload on the pool will have failed, but in
dCache an upload failure is still considered a valid file and the file is
attemped to be registered with the name space; this registration fails because
the SRM has already deleted the name space entry. The error code from this
failure is however not propagated to the pool as the pool reports the primary
failure (the FTP upload failure) rather than the post processing failure. Thus
the space manager marks the file as being stored, yet the name space entry is
gone with no name space notification reporting the file as deleted.

The second scenario triggering the same issue is if the FTP upload fails, but
the name space entry registration fails for other reasons, eg, a communication
issue with PnfsManager. The pool will mark the file broken and will eventually
repeat the registration; if the name space entry was deleted by the time this
happens, the pool will delete the file and no notification will reach space
manager.

The first scenario could be fixed by letting the post processing failure mask
the primary failure (even though that would be bad style), but it would not
at all help with the second scenario.

The issue does not exist in 2.13+ as the notification mechanism has been
changed and made independent of an actual pool location being registered.

Modification:

Space manager now doesn't immediately mark a file as stored upon encountering a
failed upload. Instead the file is queued for a recovery procedure. During
recovery the name space entry is queried; if the file was deleted the
reservation entry is removed; if a location was registered the file is marked
as stored. Otherwise the existing recovery mechanism will kick in after 24
hours and repeat this procedure.

Result:

Fixes a race condition leading to leaked space manager entries. The space
manager needs to be upgraded to resolve the issue.

Target: 2.12
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8519/
(cherry picked from commit e53a6d55dd8354329ea0576e737687ba3fa94ff8)