chimera: Refactor transaction template use
Motivation:

Spring provides a transaction template for easy interaction with the
transaction manager - in case one doesn't want the code to rely
on the transactional annotation support. The template is however
rather limited in how it deals with exceptions. We used a little
hack to work around those limitations, but it is easier to simply
not use the template at all for the JdbcFs case.

Modification:

Inlines the relevant code of TransactionTemplate and adjusts it for our
use case.

Result:

Simpler code and less overhead for propagating exceptions.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8520/