admin: Add pool group patterns
Motivation:

A common use of pcells is to reconfigure all pools in a pool group by sending a
command to the entire set. The new admin shell supports sending commands to
groups of cells, but it doesn't allow support pool groups as a source of this
list.

Modification:

Add the 'pool/group' syntax to all backslash commands. Globs are supported on
either side of the slash. An empty string on the left side is equivalent to an
asterix wildcard - an empty string on the right side has a special meaning
in that the left side will be applied to the list of all pools rather than
just the pools in a pool group (i.e. pools not in any pool group are matched
too). The expression '/' thus matches all pools in the system.

Result:

A new 'pool/group' cell list pattern. Note that it will match offline pools too
if those are listed by the pool manager.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8531/