srm: Remove third party copy from SRM front-end code
Motivation:

The SRM service is split into a front-end and a back-end. Third party
copy was implemented in several ways:

- The front-end could call out to script

- The front-end could call into the backend

- The front-end could itself copy between files and/or JVM URL streams.

Modification:

Remove the script callout and the copy between files and/or URL streams
from the front-end. The script callout code is moved to the unused
UnixFs backend, but not used.

As for dCache, only calls into the backend are supported for third party
copy.

The script callout has not been exposed in our configuration for quite
a while, so is unlikely to be in use anywhere.

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Request: 2.6
Require-notes: yes
Require-book: no

Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
(cherry picked from commit fa95576a809b385a71b5106efb651afc9e527ce6)