srm: Refactor Scheduler#schedule method
Motivation:

The Scheduler#schedule has had a somewhat ambigious role, where it
was both used to queue a request and to submit a request for
execution, depending on the state.

Modification:

Split the method into a queue method and an execute method. queue
puts the request into the QUEUED state and places it on the queue,
while EXECUTE doesn't affect the state and submits the request
for execution on the thread pool.

Calling code is updated accordingly. This also affects that code
to restore files from the database. In particular the RESTORED
state is never used as a state for requests now - it is used
as a virtual source state for the state change notification
issued when restoring a request from the database upon restart.

Result:

Minor observable differences in the event history of requests
during restart.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8552/