srm: Unpin files when aborting prepareToGet
Motivation:

The bring-online request handling has logic to abort a pin when
the request is aborted. The prepareToGet code is laking this logic.

Modification:

The primary difference is to call a catch-all unpin by ID and SURL
if the request fails before the pin has completed. This is a fairly
expensive call, but is only invoked in case of failures.

Also restructured the code slightly to make the two request implementations
more alike.

Result:

No lingering pins.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8557/