srm: Delete upload directory if SRM request gets garbage collected
Motivation:

On srmPrepareToPut a temporary upload directory gets created. Once completed,
a callback is called. It may happen that the request was aborted in the mean
time and that the request is no longer known to the SRM (garbage collected
in the database or if database persistence has been disabled entirely). In
such cases we would log an error and leave the temporary upload directory.

Modification:

Delete the upload directory if the SRM request is not retrivable. This requires
storing some extra information in the callback object.

Result:

Solves a rare race in which upload directories would leak.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8558/