dcap: Fix socket factory argument parsing
Motivation:

DCAP has a pluggable socket factory system. The argument parsing
relies on a comma separated list of arguments, but this system fails
if any of the arguments themselves contain a comma. This is the case
for our list of blocked ciphers, which means only the first in the
list is actually passed on to the socket factory.

Modification:

Use space separated arguments and make use of the Args class to get
proper quoted of arguments.

Result:

Respect all ciphers flags specified in the configuration.

Target: trunk
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Christian Bernardt <christian.bernardt@desy.de>
Patch: https://rb.dcache.org/r/8580/