build: Fix build performance regression caused by git describe
Motivation:

Our build has slowly become slower. One cause of this is the use
of git describe which adds approx. 4-5 seconds in total as the
plugin is invoked for every sub module.

Modification:

Switch to the git commit id plugin and configure it to only
run for the parent pom. The result is inherited by all other
modules.

Drop the use of the buildnumber plugin as it served a similar
purpose (to describe the build).

Dropped the Version#state property as it served a similar purpose
as the build number property.

Result:

The revision reported in webadmin/httpd is no longer a timestamp
but the git describe output.

DEBs and RPMs for SNAPSHOT builds have a version number that
includes the git describe output.

Minimum Maven version is now 3.1.1.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8590/