webdav: update file icon(s) in directory HTML page so they show file locality
A recent patch allowed the WebDAV door to know a file's locality. This
patch adds a visual representation of this information, using the free
Font-Awesome icon-set.

Rather than showing icons that represent storage media (i.e., HDD and
tape), this patch uses icons that attempt to convey the relative
delivery speeds.

Files available with ONLINE latency are shown with the normal file
icon; those with NEARLINE are shown with a truck symbol.

A file with ONLINE_AND_NEARLINE locality is shown with both file
and truck icons.

Files with LOST locality are shown with an exclamation mark in a
circle.  However, currently dCache never reports that files have
this locality.

Files with UNAVAILABLE locality are shown with a symbol similar to the
no-entry traffic sign: a circle with a horizontal line through it.

Directories continue to show their directory icon and other directory
items (such as symbolic links) are shown with the file icon, as before.

Each icon or two-icon-group include a short textual explanation.  In
most browsers this is shown when the mouse pointer hovers over the
icon(s).

Target: master
Requires-notes: yes
Requires-book: yes
Patch: https://rb.dcache.org/r/8636/
Acked-by: Gerd Behrmann
Acked-by: Femi Adeyemi