xacml: fix voms attribute validation
In the move of the XACML plugin from gPlazma 1 to 2, it was decided to continue supporting voms attribute validation as optional,
because some of the Fermi installations did not have the vomsdir properly set up.

In subsequent patches, this option, along with the configuration of the validator based on the CADIR and VOMSDIR
properties, got swallowed somehow.

As it turns out, we have actually been running validation all the time, with the default settings.

In light of this, we have decided to drop the option and enforce attribute validation (as it should be done),
and to configure the validator (as it was once done) from the paths.

We have also made the XACML path properties independent of the VOMS properties.

There is a potential here to pare down the XACML authentication routine such that it depends on the
X509 and/or VOMS principals already being in the chain, thus eliminating code duplication.  This
will be for a subsequent patch.

Target: master
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Acked-by: Gerd Behrmann
Require-book: no
Require-notes: yes