pool: Turn remote HTTP movers into dedicates transfer service and port to CANL
Motivation:

To properly inject configuration into the remote HTTP movers, reuse the PRNG
and trust managers, and to get rid of JGlobus.

Modification:

Introduces a transfer service to act as a factory and executor for remote HTTP
movers. The transfer service can maintain configuration and state between
transfers and thus allows the the PRNG and trust managers to be cached.

The code is ported to CANL.

Result:

Introduces configuration properties for CRL and OCSP checking, as well as the
CA certificate refresh interval.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8601/