ftp-client: Refactor control channel implementation
Motivation:

The GridFTPControlChannel extends FTPControlChannel by adding
GSI encryption capabilities. In the protocol the GSI encryption
is layered on top of plain unencrypted FTP, however this structure
is not apparent in the implementation. This needlessly complicates
the code.

Modification:

Changes GridFTPControlChannel to wrap a plain unencrypted control channel. The
wrapped control channel is created the moment the client authenticates with the
server. Thus the GridFTPControlChannel no longer has to switch modes, and more
of it's fields can be final.

More over, rather than duplicating command and reply parsing in the input
and output stream wrappers used by the old code, we now rely on the regular
unencrypted control channel to transport commands and replies to and from
the server.

Result:

No user visible changes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8640/