pool: Extend script driver for polling scripts
Motivation:

The classic script based interface to HSMs suffers from limited concurrency
if the scripts block until the HSM interaction has finished. Some sites
figured out that if they allow the script to fail while asynchronously starting
the HSM interaction, dCache would patiently retry the request. Internally
dCache would however treat the invocation as an error, with all the noise
that this entails.

Since the polling behaviour is really a perculiarity of the script HSM
system, the retry ought to be handled in the script driver, allowing us
to distinguish a deliberate poll request from a failure.

Modification:

Extends the script driver to recognize 72 as a special exit code requesting
that the script is invoked again later. dCache will not treat this an
error.

The script driver will delay the retry by a configurable polling period. The
new driver option -p:delay sets the number of seconds to delay between
polls. The default is 1 minute.

The poll is still subject to the concurrency settings. I.e. once poll period
expired, the request is reinserted in the request queue and at most the
configured number of requests are processed concurrently.

Result:

72 is now a special exit code causing the script driver to call the script
again after a poll period. The poll period is configurable using the new
-p:delay driver option.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/8645/