srm: Fix regression in detection of local SURLs
Motivation:

An SRM third party copy using srmCopy must either have the local SRM as
source or destination; quoting the SRM 2.2 spec:

"This function is used to copy files from source storage sites into the target
storage sites. The source storage site or the target storage site needs to be
the SRM itself that the client makes the srmCopy request to. If both source and
target are local to the SRM, it performed a local copy."

Commit ea23dc8 introduced a regression triggered when clients don't include a
port number in the SURL.

Furthermore the SRM has a flaw in that it considers any URL pointing to a host
on which the dCache has a door as being local - regardless of protocol and
regardless of port number. This is clearly against the spec as the spec says
wither source or destination has to be the SRM itself.

Modification:

- Fix the regression triggered by missing ports in the SURL by specifically
  testing for a missing port. If the port is missing we assume the SURL refers
  to the local SRM if the host name matches.

- Changes the interpretation of non-SRM sources or destinations such that they
  are never considered "local" - i.e. either source or destination MUST be the
  local SRM.

- Changes the algorithm to detect whether the host name in the SURL is the
  host running the SRM: Before a DNS lookup was made and the IP addresses
  where compared. Now the DNS lookup is skipped and the host part of the
  SURL must either be empty or match one of the detected or configured host
  names exactly (ignoring case as DNS names are not case sensitive).

Result:

- Fixes the regression.

- Limits third party copy to those having the local SRM as source or
  destination.

- Changes the detection of local SURLs to no longer resolve host names
  to IP addresses before comparison. This in particular has influences
  sites that have DNS aliases for their SRM and they may have to add
  those aliases to the list of local names in srm.net.local-hosts.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8671/