voms: Fix resource leak regression
Motivation:

The voms validator starts an internal thread to periodically reload
the vomsdir configuration. Because of this thread it is crucial to
reuse the voms trust store as well as to properly shut it down when
not using it anymore. Most certain we should not instantiate the
trust store per connection as we currently do in gsidcap and for
gridsite delegation.

Modification:

Instantiate the VOMS validator once per service. Where possible we
shut it down when the service is shut down (this is not always
possible due to interface limitations).

For grid site delegation this patch also makes it respect the vomsdir
and capath configuration of the srm service.

Result:

Avoid resource leak in gsidcap and srm. Make srm respect capath and
vomsdir configuration.

Target: 2.13
Require-notes: yes
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>
Patch: https://rb.dcache.org/r/8682/