srm: Replace AuthorizationRecord with new user manager
Motivation:

AuthorizationRecord is a complicated structure using JPA for persistence.  It
was originally intended as a generic format to represent users in dCache, but
now it is merely used for persisting user identities in the SRM database.

AuthorizationRecord has several issues, the most important being that it
doesn't accurately capture the LoginReply generated by gPlazma. Since gPlazma
is pluggable, it may include arbitrary principals, yet AuthorizationRecord only
maps a handful of principals and login attributes to its internal format.

Moreover, it represents X.500 DNs in string form. It is known that strings
cannot accurately capture X.500 DNs, meaning that the mapping back to an
X500Principal is incomplete. This hasn't been a problem with GlobusPrincipal as
it itself is text based, however since we want to move away from
GlobusPrincipal it will become a problem.

Finally, AuthorizationRecord doesn't use liquibase schema management and is the
only dependency in dCache on JPA.

There are several alternatives for persisting the user identity in the SRM:

- Persisting the LoginReply to the database. This was previously done by
  converting to AuthorizationRecord and back. Without using AuthorizationRecord
  we need to be able to persist an abitrary collection of principals and
  login attributes. Conversion to a string form is incomplete for X.500
  principals as well as group principals (information about primary groups
  is lost). Since gPlazma is pluggable, there may be other principals or
  attributes we don't know yet. Thus a solution based on writing LoginReply
  to the database has to rely on Java's generic serialization support.

- Persisting the certificate chain of the user. Upon restoring it from the
  database it can be mapped to a LoginReply by "logging into" gPlazma. With
  caching of gPlazma replies, this solution performs well too and avoids
  the user of Java serialization for persistence.

The problem with the former is the user of Java serialization for persistence.
The problem of the latter is that gPlazma configuration may have changed and
thus the user may no longer be authorized when restoring the chain from the
database. This may actually be considered a feature.

Modification:

The patch implements both solutions described above, although only the
one serializing the LoginReply is currently used.

A CanonicalizingByteArrayStore assigns unique identifiers to the encoded user
(the encoding being either the encoded certificate path or serialized
principals and login attributes) using hashing with rehashing to resolve
collisions - this is just like what AuthorizationRecord did, although the hash
is obviously different. CanonicalizingByteArrayStore also provides caching and
a means of deleting unreferenced users in the SRM database.

DcacheUserManager is a base class for user managers using
CanonicalizingByteArrayStore with a JDBC backend and uses DcacheUser to
represent users. Two subclasses implement the two different encodings described
above.

The schema is managed by liquibase. Upon upgrade the old AuthorizationRecord
tables are deleted and the userid column of the request tables is nulled. This
means that upon upgrade existing requests loose their user information. To
allow such requests to at least fail gracefully and thus unpin files or delete
upload directories, the user manager can generate anonymous user identities in
this case. Storage will refuse to perform any operation for these anonymous
users. The failure handling code for requests is updated to unpin/abort uploads
using a root identity rather than the owner of the request (there is no point
in those cleanup operations failing with authorization errors when the SRM
request already failed).

Anonymous users are also used when loading a certificate chain that fails
mapping in gPlazma. Note that these are only ever used as a substitute for a
user identity that used to be authorized. The primary purpose is to allow
the request to be loaded from the database and let the SRM at least abort
it.

Result:

Liquibase managed user persistence in the SRM which accurately captures the
full gPlazma login. Upon upgrade, user records of existing requests are
lost and inprogress requests will fail.

I hope this can make it into 2.14 as it would mean that SRM is fully managed
by liquibase.

Target: trunk
Request: 2.14
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8688/