srm: Drop indexes on nextjobid column
Motivation:

We don't use the column and an index consumes space and costs time
to maintain.

Modification:

Remove the indexes on the nextjobid column.

Result:

The schema is updated to drop the indexes. Rollback actions are
defined to rollback to a schema compatible with the previous
release.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8690/