Logging CANL notifications to correct cells context
Motivation:

CANL uses an internal static timer to trigger reload of trust store material.
Because of this, our listener gets called with whatever logging context the
thread used for this timer was created within.

Modification:

Since CDC is in the cells module and since common-security doesn't depend on cells,
we cannot directly restore the CDC in the listener. Instead a supplier of logging
contexts is added to the CanlContextFactory builder. Thus the caller of the builder
can inject a supplier based on the CDC class.

Result:

Notifications are logged to the correct cell.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8722/