Fix shutdown of bounded executor
Motivation:

BoundedExecutor wraps another executor to provide a bound on the number of threads
being used. It does not assume control over the lifecycle of the wrapped executor
as one of the use cases is to have several bounded executors source threads from
the same executor.

Yet for convenience the class provides constructors to create and wrap a cached
thread pool. In that case the wrapped executor is never shut down.

Modification:

Remove the two constructors that create a cached thread pool. Instead a subclass
is created to provide this functionality. The subclass terminates the wrapped
executor on shutdown.

Result:

No hanging thread pool on shutdown.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.14
Request: 2.13
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8738/