cells: Make cells message executor configurable
Motivation:

Currently we have a mechanism for injecting a custom executor into the cell
nucleus to be used for message processing. This is used both to tune the
concurrency/queue limits of the cell, as well as to inject shared executors
to use accross several cells.

The problem with the current design is that it presents a problem with how to
initialize the custom executor: Initialization should be done on the message
executor to get the correct thread group and CDC inheritance, but that means
a custom executor should be created before regular initialization or we have
to first start on executor for initialization and then replace it with another
one. The latter is currently done for UniversalSpringCell.

Modification:

With this patch a bounded executor is used for message delivery and cannot
be replaced. The thread and queue limits are however configurable through
cell arguments. Since a bounded executor wraps another executor, a shared
executor can be realized by letting multiple cells wrap the same executor
(this is done by the line based door).

The cell message executor is exposed to the spring configuration to allow
cell components to execute tasks on it.

Result:

Significantly simpler message executor design and no duplicate executor
creation during startup. The configuration properties for setting the limits
have been harmonized and the old names have been deprecated.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8752/