pom: fix dirty version tag for deb and rpm
c1d2fe151 changed the version tag for dirty versions to use
underscored as separators to fix an issue with rpm, but it
introduced an issue for dpkg-buildpackge that caused deb
builds to fail for dirty versions.

This patch changes the separator to "+" which works for both.

Ticket:
Acked-by:
Target: trunk
Require-book: no
Require-notes: no