cells: Don't use CellExceptionMessage
Motivation:

If we are to replace the wire protocol for domain to domain communication, we
cannot allow CellMessage to be arbitrarily subclassed. The best is to mark the
class final, but to get there we have to get rid of the CellExceptionMessage
subclass first.

CellExceptionMessage is used to carry NoRouteToCellException and allows
CellGlue to detect no-route replies to avoid that they bounce between domains.

Modification:

The use cases in the PnfsManager and CellAdapter are replies by direct replies
using the original envelope: These cases already detect replies and only reply
to request messages.

LocationMgrTunnel is updated to use the original envelope too; the cell only
generates a no route to cell exception before committing suicide and thus there
is no risk of bounces.

When CellGlue generates a no route to cell exception is includes the domain
address as the source. This provides a convenient marker as a domain itself
cannot receive any messages anyway; thus if a domain address (one using an
asterix for the cell name) is detected, there is no point in generating an
error response for it. In turn, CellGlue can too use the original envelope when
generating a response.

The CellExceptionMessage class is kept for compatibility with pre-2.15 domains.
It suffices that we can deserialize objects of this class: Two pre-2.15 domains
sending CellExceptionMessage through a 2.15 domain will receive the
CellExceptionMessage as is; a 2.15 domain sending a no route to cell exception
with a domain source to a pre-2.15 domain will possibly receive a
CellExceptionMessage bounce back, but a 2.15 domain receiving a
CellExceptionMessage from a pre-2.15 domain will detect the domain source and
not itself generating a no route to cell response in case of delivery failure;
thus even between mixed version domains there is no risk of bouncing delivery
failures.

It is critical for pre-2.15 compatibility that CellExceptionMessage is forwarded
between such domains as both rely on this to suppress bounces. This has to be
kept in mind when evolving the inter-doain protocol.

Result:

CellExceptionMessage is not used directly, but needs to be kept for pre-2.15
compatibility. It can be dropped after the next golden release.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8783/