(2.14) alarms: modify logback encoding to show actual alarm type
Motivation:

Logback pattern encoding allows for the display of marker names
using either '%marker' or '%markerSimpleName'.  The way Alarms currently
uses markers, however, is to chain the actual value of a sub-marker as
another sub-marker.

The current encoding thus merely displays the marker names, which is
not all that informative (and perhaps even annoying): e.g.,

17 Nov 2015 07:07:17 ALARM [ ALARM_TYPE ] (fndca3a.fnal.gov)(user-command)(<na>) This is a test alarm for OutOfMemory error reporting.

Modification:

Upon arrival of the event at the server, during the preprocessing of
the alarm for either email or the optional history file, the submarker
is accessed and added to the event's MDC property map.  The pattern
encoding is changed to display the new MDC property instead of the marker.

Result:

A more meaningful printout, e.g.,

17 Nov 2015 07:07:17 FATAL_JVM_ERROR (fndca3a.fnal.gov)(user-command)(<na>) This is a test alarm for OutOfMemory error reporting.

Target: 2.14
Require-book: no
Require-notes: yes
Acked-by: Gerd
Patch: https://rb.dcache.org/r/8781/